Addition of harvester
Since the harvester was add we removed two of the lift motors to use
for the harvester, but we might only use one.  So instead of useing
both motorF and motorG for the lift, we are going to only use motorF
for the lift, and use motorG for the harvester.  The harvester is to be
controlled by the second joystick on the second cotroller so that we
could have variable speeds.