Autonomous Revision
We are now having the autonomous run backwards so that we are able to
score the cube on the left.  We are hard coding for this regionals and
there is plans for adding sensors in the future: Gyro, compass, light,
sonar.  We were able to sucsessfully score in all baskets.  Turns have
been perfected, and compensations have been added to stay on course.
Move method has been added so that testing autonomous is easier.
Comments were also added to main task to describe actions.